+++ src/Learner.php
@@ -2,24 +2,8 @@
 
 namespace Rubix\ML;
 
-use Rubix\ML\Datasets\Dataset;
-
-interface Learner extends Estimator
+interface Learner extends Trainable, Estimator
 {
-    /**
-     * Train the learner with a dataset.
-     *
-     * @param \Rubix\ML\Datasets\Dataset $dataset
-     */
-    public function train(Dataset $dataset) : void;
-
-    /**
-     * Has the learner been trained?
-     *
-     * @return bool
-     */
-    public function trained() : bool;
-
     /**
      * Predict a single sample and return the result.
      *