+++ src/RanksFeatures.php
@@ -4,7 +4,7 @@ namespace Rubix\ML;
 
 use Stringable;
 
-interface RanksFeatures extends Stringable
+interface RanksFeatures extends Trainable, Stringable
 {
     /**
      * Return the normalized importance scores of each feature column of the training set.